core/google_sheets_service.py
from __future__ import print_function

import os.path

from google.auth.transport.requests import Request
from google.oauth2.credentials import Credentials
from google_auth_oauthlib.flow import InstalledAppFlow
from googleapiclient.discovery import build
from googleapiclient.errors import HttpError

# If modifying these scopes, delete the file token.json.
SCOPES = ['https://www.googleapis.com/auth/spreadsheets']

_global_sheet_service = None


# creates a new sheet given name, returns spreadsheetId
def create_sheet(sheet_name: str) -> str:
    sheet = get_sheet()

    return sheet.create(
        body={"properties": {"title": sheet_name}},
        fields="spreadsheetId"
    ).execute()["spreadsheetId"]


def get_sheet():
    global _global_sheet_service
    if _global_sheet_service is None:
        _global_sheet_service = get_google_sheet_service().spreadsheets()
    sheet = _global_sheet_service
    return sheet


def get_google_sheet_service():
    """Shows basic usage of the Sheets API.
    Prints values from a sample spreadsheet.
    """
    creds = None
    # The file token.json stores the user's access and refresh tokens, and is
    # created automatically when the authorization flow completes for the first
    # time.
    if os.path.exists('token.json'):
        creds = Credentials.from_authorized_user_file('token.json', SCOPES)
    # If there are no (valid) credentials available, let the user log in.
    if not creds or not creds.valid:
        if creds and creds.expired and creds.refresh_token:
            creds.refresh(Request())
        else:
            flow = InstalledAppFlow.from_client_secrets_file(
                'credentials.json', SCOPES)
            creds = flow.run_local_server(port=0)
        # Save the credentials for the next run
        with open('token.json', 'w') as token:
            token.write(creds.to_json())

    try:
        service = build('sheets', 'v4', credentials=creds)
        return service
    except HttpError as err:
        print(err)
