requirements.txt
ipython==8.9.0
dill==0.3.6
apscheduler==3.9.1.post1
datetime==5.0
requests==2.28.2
google-api-core==2.11.0
google-api-python-client==2.74.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.58.0
pandas==1.5.3
numpy==1.24.1
pyfzf
matplotlib